fix(skills): misc skill fixes around health & morale
